move left over operators/sensors tests inside standard provider
