+++ .pre-commit-config.yaml
@@ -724,7 +724,10 @@ repos:
         files: >
           (?x)
           ^providers/src/airflow/providers/.*\.py$
-        exclude: ^.*/.*_vendor/|providers/src/airflow/providers/standard/operators/bash.py|providers/src/airflow/providers/standard/operators/python.py
+        exclude: |
+          (?x)
+          ^.*/.*_vendor/|
+          ^providers/src/airflow/providers/standard/.*$
       - id: check-get-lineage-collector-providers
         language: python
         name: Check providers import hook lineage code from compat
