+++ airflow/example_dags/example_sensors.py
@@ -24,11 +24,11 @@ import pendulum
 from airflow.models.dag import DAG
 from airflow.providers.standard.operators.bash import BashOperator
 from airflow.providers.standard.sensors.bash import BashSensor
+from airflow.providers.standard.sensors.filesystem import FileSensor
 from airflow.providers.standard.sensors.python import PythonSensor
 from airflow.providers.standard.sensors.time import TimeSensor, TimeSensorAsync
 from airflow.providers.standard.sensors.time_delta import TimeDeltaSensor, TimeDeltaSensorAsync
 from airflow.providers.standard.sensors.weekday import DayOfWeekSensor
-from airflow.sensors.filesystem import FileSensor
 from airflow.utils.trigger_rule import TriggerRule
 from airflow.utils.weekday import WeekDay
