+++ providers/src/airflow/providers/standard/operators/datetime.py
@@ -20,7 +20,7 @@ import datetime
 from typing import TYPE_CHECKING, Iterable
 
 from airflow.exceptions import AirflowException
-from airflow.operators.branch import BaseBranchOperator
+from airflow.providers.standard.operators.branch import BaseBranchOperator
 from airflow.utils import timezone
 
 if TYPE_CHECKING:
