+++ providers/tests/standard/operators/test_branch_operator.py
@@ -22,8 +22,8 @@ import datetime
 import pytest
 
 from airflow.models.taskinstance import TaskInstance as TI
-from airflow.operators.branch import BaseBranchOperator
 from airflow.operators.empty import EmptyOperator
+from airflow.providers.standard.operators.branch import BaseBranchOperator
 from airflow.utils import timezone
 from airflow.utils.state import State
 from airflow.utils.task_group import TaskGroup
