+++ providers/tests/standard/sensors/test_external_task_sensor.py
@@ -38,12 +38,12 @@ from airflow.models.xcom_arg import XComArg
 from airflow.operators.empty import EmptyOperator
 from airflow.providers.standard.operators.bash import BashOperator
 from airflow.providers.standard.operators.python import PythonOperator
-from airflow.providers.standard.sensors.time import TimeSensor
-from airflow.providers.standard.triggers.external_task import WorkflowTrigger
 from airflow.providers.standard.sensors.external_task import (
     ExternalTaskMarker,
     ExternalTaskSensor,
 )
+from airflow.providers.standard.sensors.time import TimeSensor
+from airflow.providers.standard.triggers.external_task import WorkflowTrigger
 from airflow.serialization.serialized_objects import SerializedBaseOperator
 from airflow.utils.hashlib_wrapper import md5
 from airflow.utils.session import create_session, provide_session
@@ -934,8 +934,8 @@ exit 0
             ),
         ),
     )
-    @mock.patch("airflow.sensors.external_task.ExternalTaskSensor.get_count")
-    @mock.patch("airflow.sensors.external_task.ExternalTaskSensor._get_dttm_filter")
+    @mock.patch("airflow.providers.standard.sensors.external_task.ExternalTaskSensor.get_count")
+    @mock.patch("airflow.providers.standard.sensors.external_task.ExternalTaskSensor._get_dttm_filter")
     def test_fail_poke(
         self, _get_dttm_filter, get_count, soft_fail, expected_exception, kwargs, expected_message
     ):
@@ -991,7 +991,7 @@ exit 0
             ),
         ),
     )
-    @mock.patch("airflow.sensors.external_task.ExternalTaskSensor._get_dttm_filter")
+    @mock.patch("airflow.providers.standard.sensors.external_task.ExternalTaskSensor._get_dttm_filter")
     @mock.patch("airflow.models.dagbag.DagBag.get_dag")
     @mock.patch("os.path.exists")
     @mock.patch("airflow.models.dag.DagModel.get_current")
