+++ providers/tests/standard/sensors/test_filesystem.py
@@ -26,8 +26,8 @@ import pytest
 
 from airflow.exceptions import AirflowSensorTimeout, TaskDeferred
 from airflow.models.dag import DAG
+from airflow.providers.standard.sensors.filesystem import FileSensor
 from airflow.providers.standard.triggers.file import FileTrigger
-from airflow.sensors.filesystem import FileSensor
 from airflow.utils.timezone import datetime
 
 pytestmark = pytest.mark.db_test
