+++ tests/serialization/test_dag_serialization.py
@@ -2241,7 +2241,7 @@ class TestStringifiedDAGs:
 
         assert tasks[0]["__var"]["start_trigger_args"] == {
             "__type": "START_TRIGGER_ARGS",
-            "trigger_cls": "airflow.triggers.temporal.TimeDeltaTrigger",
+            "trigger_cls": "airflow.providers.standard.triggers.temporal.TimeDeltaTrigger",
             # "trigger_kwargs": {"__type": "dict", "__var": {"delta": {"__type": "timedelta", "__var": 2.0}}},
             "trigger_kwargs": {"__type": "dict", "__var": {"delta": {"__type": "timedelta", "__var": 2.0}}},
             "next_method": "execute_complete",
